+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # Changelog 
+## Version 0.62.9
+- added an option to allow the overview chart to start above 0°
+- modified the possible values for a fixed scale to allow some values above 0°
+- added an error check to prevent min>max when a fixed scale is enabled
+- WidgetDimensionManager guesses a widget size based on available window size when the reported widget dimensions are implausible
+- improved font size of the clock in the clock widget
+
 ## Version 0.62.8
 - fixed stations history displaying "null" when warnings are disabled and municipality names instead of the station name are displayed
 - remove pre-fetching of map images to speed up main ui on older devices